discarding series_generators and moving important functions to utils
